Clarify purpose of mining address
Fixes #38

This change clarifies the role of the mining address. We emphasize that
the address is public and belongs to some account. We only hint at
generating a key pair and reference other documentation instead.

If that makes it clearer we could also explicitly state that there is no
need to run `radicle-registry-cli` on the same machine as
`radicle-registry-node` and that they do not share data.